requirements.txt
bcrypt==3.2.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
python-dotenv==0.14.0
PyJWT==1.7.1
